지원자 공고화면개선 진행 251204-1

+++ lib/screens/business_admin/dialogs/confirmed_list_dialog.dart
@@ -673,7 +673,7 @@ class _ConfirmedListDialogWidgetState
         widget.toItem.workDetailStats ??= {};
         final stats = widget.toItem.workDetailStats![workType];
         if (stats != null) {
-          stats['confirmed'] = ((stats['confirmed'] ?? 1) as int) - 1;
+          stats['confirmed'] = ((stats['confirmed'] ?? 1)) - 1;
         }
       });
     } else {
